# Modify _encode_datetime_with_cftime for compatibility with cftime > 1.4.0