--- a/doc/whats-new.rst
+++ b/doc/whats-new.rst
@@ -97,7 +97,7 @@ Bug fixes
   By `Julien Seguinot <https://github.com/juseg>`_.
 - Add :py:meth:`Dataset.drop_isel` and :py:meth:`DataArray.drop_isel` (:issue:`4658`, :pull:`4819`). By `Daniel Mesejo <https://github.com/mesejo>`_.
 - Fix time encoding bug associated with using cftime versions greater than
-  1.4.0 with xarray (:issue:`4870`). By `Spencer Clark <https://github.com/spencerkclark>`_.
+  1.4.0 with xarray (:issue:`4870`, :pull:`4871`). By `Spencer Clark <https://github.com/spencerkclark>`_.
 
 Documentation
 ~~~~~~~~~~~~~