--- a/xarray/tests/test_coding_times.py
+++ b/xarray/tests/test_coding_times.py
@@ -12,6 +12,7 @@ from xarray.coding.times import (
     decode_cf_datetime,
     encode_cf_datetime,
     to_timedelta_unboxed,
+    _encode_datetime_with_cftime
 )
 from xarray.coding.variables import SerializationWarning
 from xarray.conventions import _update_bounds_attributes, cf_encoder
@@ -995,3 +996,18 @@ def test_encode_decode_roundtrip(freq):
     encoded = conventions.encode_cf_variable(variable)
     decoded = conventions.decode_cf_variable("time", encoded)
     assert_equal(variable, decoded)
+
+
+@requires_cftime
+def test__encode_datetime_with_cftime():
+    # See GH 4870. cftime versions > 1.4.0 required us to adapt the
+    # way _encode_datetime_with_cftime was written.
+    import cftime
+
+    calendar = "gregorian"
+    times = cftime.num2date([0, 1], "hours since 2000-01-01", calendar)
+
+    encoding_units = "days since 2000-01-01"
+    expected = cftime.date2num(times, encoding_units, calendar)
+    result = _encode_datetime_with_cftime(times, encoding_units, calendar)
+    np.testing.assert_equal(result, expected)