@mathause 2021-02-07
LGTM - did you check if that makes it slower? (Probably not as `vectorize` is also a for loop under to hood).

I think the failures are #4873 and #4865