@spencerkclark 2021-02-07
Thanks @mathause -- indeed the timings are pretty similar. 

Before:
```
In [1]: import numpy as np; import xarray as xr

In [2]: times = xr.cftime_range("2000", periods=1000).values

In [3]: units = "days since 2000-01-01"

In [4]: calendar = "gregorian"

In [5]: %%timeit
   ...: xr.coding.times._encode_datetime_with_cftime(times, units, calendar)

15.5 ms ± 170 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After:
```
In [1]: import numpy as np; import xarray as xr

In [2]: times = xr.cftime_range("2000", periods=1000).values

In [3]: units = "days since 2000-01-01"

In [4]: calendar = "gregorian"

In [5]: %%timeit
   ...: xr.coding.times._encode_datetime_with_cftime(times, units, calendar)

14.9 ms ± 160 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```